Separate legacy code out of the app.ts file
