+++ test/users.test.ts
@@ -34,4 +34,8 @@ describe("Users tests", () => {
       });
     });
   });
+
+  describe("CreateUser tests", () => {
+    it("should be able to create a new user", async () => {});
+  });
 });
